Add convenient methods for Image (#10221)
# Objective
To get the width or height of an image you do:
```rust
self.texture_descriptor.size.{width, height}
```
that is quite verbose.
This PR adds some convenient methods for Image to reduce verbosity.

## Changelog
* Add a `width()` method for getting the width of an image.
* Add a `height()` method for getting the height of an image.
* Rename the `size()` method to `size_f32()`.
* Add a `size()` method for getting the size of an image as u32.
* Renamed the `aspect_2d()` method to `aspect_ratio()`.

## Migration Guide
Replace calls to the `Image::size()` method with `size_f32()`.
Replace calls to the `Image::aspect_2d()` method with `aspect_ratio()`.

+++ crates/bevy_render/src/texture/compressed_image_saver.rs
@@ -40,7 +40,7 @@ impl AssetSaver for CompressedImageSaver {
 
         let mut source_image = compressor_params.source_image_mut(0);
         let size = image.size();
-        source_image.init(&image.data, size.x as u32, size.y as u32, 4);
+        source_image.init(&image.data, size.x, size.y, 4);
 
         let mut compressor = basis_universal::Compressor::new(4);
         // SAFETY: the CompressorParams are "valid" to the best of our knowledge. The basis-universal
